Refactor to pass page as integer in user table rendering
Updated the userTable method in the UserController to convert the 'page' path variable to an integer before passing it to the 'render' method. This change was done to make the 'render' method in the UserTableComponent accept an integer for page number, providing type safety. Adjusted the position of the getPage method in the UserService class as well.